Revert "add react-native-bouncy-checkbox dependency"
This reverts commit e2bd5267fa1d18e7a8de16e8ebbfb076b874ba84.
