+++ PET_ME/android/app/src/main/res/drawable/image_svg_2cdc2c.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="55dp"
+    android:height="54dp"
+    android:viewportWidth="55"
+    android:viewportHeight="54">
+  <path
+      android:pathData="M15.9432,53.6667H39.0542C48.095,53.6667 54.1667,47.325 54.1667,37.8887V16.1113C54.1667,6.675 48.095,0.3333 39.0569,0.3333H15.9432C6.9051,0.3333 0.8334,6.675 0.8334,16.1113V37.8887C0.8334,47.325 6.9051,53.6667 15.9432,53.6667ZM18.1637,24.3333C14.4873,24.3333 11.5,21.342 11.5,17.6667C11.5,13.9913 14.4873,11 18.1637,11C21.8374,11 24.8273,13.9913 24.8273,17.6667C24.8273,21.342 21.8374,24.3333 18.1637,24.3333ZM48.3555,34.8239C49.2485,37.1135 48.7846,39.8655 47.83,42.1331C46.6984,44.8301 44.5317,46.872 41.8019,47.7637C40.5898,48.16 39.3188,48.3333 38.0504,48.3333H15.5764C13.34,48.3333 11.361,47.7967 9.7387,46.7977C8.7224,46.1703 8.5428,44.7227 9.2963,43.7843C10.5566,42.2157 11.8008,40.6415 13.0558,39.0537C15.4477,36.0155 17.0593,35.1348 18.8506,35.9081C19.5773,36.2274 20.3067,36.7062 21.0575,37.2126C23.0579,38.5721 25.8387,40.4407 29.5017,38.4124C32.0083,37.0085 33.4622,34.6003 34.7282,32.5033L34.7494,32.4682C34.8391,32.3211 34.928,32.1739 35.0167,32.0272L35.0168,32.0272C35.4422,31.3232 35.862,30.6286 36.3369,29.9886C36.9322,29.1878 39.1391,26.6835 41.9976,28.4668C43.8184,29.5896 45.3495,31.1087 46.988,32.7351C47.6128,33.3571 48.0579,34.0643 48.3555,34.8239Z"
+      android:fillColor="#64868E"
+      android:fillType="evenOdd"/>
+</vector>
