+++ PET_ME/ios/PET_ME/Images.xcassets/image_svg_2cdc2c.imageset/Contents.json
@@ -1,0 +1,15 @@
+{
+  "images": [
+    {
+      "filename": "light.pdf",
+      "idiom": "universal"
+    }
+  ],
+  "info": {
+    "author": "xcode",
+    "version": 1
+  },
+  "properties": {
+    "preserves-vector-representation": true
+  }
+}
