--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "react-native-bouncy-checkbox": "^2.1.10"
-  }
-}
